Tweak milestone placement in day mode.  Refs #9443.
With this change, the milestone shows up on the border between days in
FF 7.0.1 on Ubuntu ''at default magnification''.  Zooming in with Ctrl-+,
shifts the milestone's diamond to the right.  The shift seems to be
worse as you go further to the right on the chart.

I believe this is an improvement but I need to figure out why the shift
happens.