Rename internal fields to avoid conflicts.  Refs #9648.
Custom field names must start with a letter and have only letters,
digits, and underscores.  By adding a leading underscore to "calc_start"
and "calc_finish", we avoid a potential conflict if the user used those
as custom field names for their own purposes or even for `fields.start`
and `fields.finish` values.