AccountManagerPlugin: Add tool-tip to input field for upper lock time limit.
While moving your mouse to that field, most browsers will show a more
human-readable timedelta representation of the value in seconds now.
It looks valuable to me, because that option value is getting too big for
fast checks at a glance.

Work on the underlying function has been pushed only up to the point,
where it became useful for this application.  There's room for development
and improvements - postponed until required.